[#1770] Code cleanup
