+++ akvo/iati/imports/fields/__init__.py
@@ -4,24 +4,22 @@
 # See more details in the license.txt file located at the root folder of the Akvo RSR module.
 # For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
 
-from .conditions import conditions
 from .contacts import contacts
 from .classifications import policy_markers, sectors
 from .dates import actual_end_date, actual_start_date, planned_end_date, planned_start_date
-from .defaults import (currency, collaboration_type, default_aid_type, default_finance_type,
-                       default_flow_type, default_tied_status, hierarchy, language, scope)
+from .defaults import (currency, collaboration_type, conditions, default_aid_type,
+                       default_finance_type, default_flow_type, default_tied_status, hierarchy,
+                       language, scope, status)
 from .descriptions import (background, current_status, custom_fields, goals_overview, project_plan,
                            project_plan_summary, sustainability, target_group, title, subtitle)
 from .financials import (budget_items, capital_spend, country_budget_items, planned_disbursements,
                          transactions)
-from .legacy_data import legacy_data
 from .links import current_image, documents, links
 from .locations import locations, recipient_countries, recipient_regions
 from .partnerships import partnerships
 from .related_projects import related_projects
 from .results import results
-from .special_reporting import crs_add, fss
-from .status import status
+from .special_reporting import crs_add, fss, legacy_data
 
 __all__ = [
     'actual_end_date',
